style: rename keypress to keyboard
